feat: Add userQuery function to retrieve user data by ID
This commit adds a new function called `userQuery` to the `data.js` file in the `utils` directory. The function takes a `userId` parameter and returns a query string that retrieves user data from the database based on the provided ID. This function will be used to fetch user data in various parts of the application.