feat: Update Home component to use fetchUser utility function
The Home component in the frontend is updated to use the newly added fetchUser utility function. This function retrieves the user data from the local storage and returns it as a JSON object. If no user data is found, the local storage is cleared. This enhancement improves the user management functionality of the application.